Refactored tabbedClaimSummary to have SectionsPagerAdapter as its own class
