+++ ShinyExpenseTracker/res/layout/activity_tabbed_summary.xml
@@ -1,3 +1,5 @@
+<!-- ViewPager used to swipe between tabs -->
+
 <android.support.v4.view.ViewPager xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/pager"
