+++ ShinyExpenseTracker/res/values/strings.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-
 <resources>
 
     <string name="app_name">ShinyExpenseTracker</string>
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
-
     <string name="title_activity_view_expense_claim">ViewExpenseClaimActivity</string>
     <string name="title_activity_add_expense_claim">AddExpenseClaimActivity</string>
     <string name="title_activity_manage_tag">Manage Tags</string>
@@ -51,5 +49,6 @@
     <string name="label_expense_totals">Expense Totals:</string>
     <string name="label_start_date">Start Date:</string>
     <string name="label_status">Status:</string>
+    <string name="title_activity_claim_summary">ClaimSummaryActivity</string>
 
 </resources>
