+++ ShinyExpenseTracker/src/ca/ualberta/cs/shinyexpensetracker/activities/TabbedSummaryActivity.java
@@ -1,24 +1,23 @@
 package ca.ualberta.cs.shinyexpensetracker.activities;
 
-import java.util.Locale;
-
 import android.app.ActionBar;
-import android.app.Activity;
-import android.app.Fragment;
-import android.app.FragmentManager;
 import android.app.FragmentTransaction;
+import android.content.Context;
 import android.os.Bundle;
 import android.support.v13.app.FragmentPagerAdapter;
+import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.widget.Toast;
 import ca.ualberta.cs.shinyexpensetracker.R;
+import ca.ualberta.cs.shinyexpensetracker.SectionsPagerAdapter;
 
 // Source: https://github.com/astuetz/PagerSlidingTabStrip
 // on March 11 2015
+//http://www.androidhive.info/2013/10/android-tab-layout-with-swipeable-views-1/
+//on March 12 2015
 
-public class TabbedSummaryActivity extends Activity implements
+public class TabbedSummaryActivity extends FragmentActivity implements
 		ActionBar.TabListener {
 
 	/**
@@ -34,6 +33,7 @@ public class TabbedSummaryActivity extends Activity implements
 	 * The {@link ViewPager} that will host the section contents.
 	 */
 	ViewPager mViewPager;
+	Context context;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -44,6 +44,8 @@ public class TabbedSummaryActivity extends Activity implements
 		final ActionBar actionBar = getActionBar();
 		actionBar.setNavigationMode(ActionBar.NAVIGATION_MODE_TABS);
 
+		context = getBaseContext();
+		
 		// Create the adapter that will return a fragment for each of the three
 		// primary sections of the activity.
 		mSectionsPagerAdapter = new SectionsPagerAdapter(getFragmentManager());
@@ -70,7 +72,7 @@ public class TabbedSummaryActivity extends Activity implements
 			// the TabListener interface, as the callback (listener) for when
 			// this tab is selected.
 			actionBar.addTab(actionBar.newTab()
-					.setText(mSectionsPagerAdapter.getPageTitle(i))
+					.setText(mSectionsPagerAdapter.getPageTitle(i, context))
 					.setTabListener(this));
 		}
 	}
@@ -110,54 +112,5 @@ public class TabbedSummaryActivity extends Activity implements
 	@Override
 	public void onTabReselected(ActionBar.Tab tab,
 			FragmentTransaction fragmentTransaction) {
-	}
-
-	/**
-	 * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
-	 * one of the sections/tabs/pages.
-	 */
-	public class SectionsPagerAdapter extends FragmentPagerAdapter {
-		public SectionsPagerAdapter(FragmentManager fm) {
-			super(fm);
-		}
-
-		@Override
-		public Fragment getItem(int position) {
-			// getItem is called to instantiate the fragment for the given page.
-			// Return a PlaceholderFragment (defined as a static inner class
-			// below).
-			Toast.makeText(TabbedSummaryActivity.this, ""+position, Toast.LENGTH_SHORT).show();
-			switch (position) {
-			case 0:
-				return ClaimSummaryFragment.newInstance(position+1);
-			case 1:
-				return ExpenseItemListFragment.newInstance(position+1);
-			case 2:
-				return DestinationsListFragment.newInstance(position+1);
-			default:
-				throw new RuntimeException("No such tab");
-			}
-		}
-
-		@Override
-		public int getCount() {
-			// Show 3 total pages.
-			return 3;
-		}
-
-		@Override
-		public CharSequence getPageTitle(int position) {
-			Locale l = Locale.getDefault();
-			switch (position) {
-			case 0:
-				return getString(R.string.title_section1).toUpperCase(l);
-			case 1:
-				return getString(R.string.title_section2).toUpperCase(l);
-			case 2:
-				return getString(R.string.title_section3).toUpperCase(l);
-			default:
-				throw new RuntimeException("No such tab");
-			}
-		}
 	}
 }
